Move OutcomeDetailsFormatter.kt to same packet as SavedMatrix.kt
